Рефакторинг кода.
